unsigned replaced by size_t

+++ src/Boundary/Boundary/GaussSolver.cpp
@@ -5,17 +5,17 @@
 
 Vector GaussSolver::Solve(UsualMatrix& a, Vector& b)
 {
-	const unsigned n = a.size1();
+	const size_t n = a.size1();
 
 	// forward elimination
-	for (unsigned j = 0; j < n - 1; ++j) {
+	for (size_t j = 0; j < n - 1; ++j) {
 		const double pivot = a(j, j);
 		if (pivot == 0)
 			throw "Gauss: zero on the main diagonal";
 
-		for (unsigned i = j + 1; i < n; ++i) {
+		for (size_t i = j + 1; i < n; ++i) {
 			const double mult = a(i, j) / pivot;
-			for (unsigned k = j; k < n; ++k)
+			for (size_t k = j; k < n; ++k)
 				a(i, k) -= a(j, k) * mult;
 			b[i] -= b[j] * mult;
 		}
